feat(AdminProfiles): Refactor to list with filter
